New uniform_other internal module
